requirements.txt
aiofiles
aiohttp
asyncio
dnspython
ffmpeg-python
gitpython
hachoir
heroku3
lyricsgenius
motor==3.0.0
pillow==9.5.0
psutil
py-tgcalls==0.9.7
pykeyboard
pyrogram==1.4.16
python-dotenv
pyyaml
requests
speedtest-cli
spotipy
tgcrypto
numpy
git+https://github.com/yt-dlp/yt-dlp
youtube-search
youtube-search-python
